refactor the folder structure of the components
