+++ src/app/(index)/page.tsx
@@ -1,5 +1,5 @@
 import {Button} from "@/components/ui/button/animated-button"
-import Dock from "@/components/app/Dock"
+import {Dock} from "@/components/app/dock"
 import Summary from "@/components/app/Summary"
 import {IonChevronForward} from "@/components/icons/ion"
 import {Container} from "@/components/layout/container"
