+++ src/components/app/dock/index.ts
@@ -1,0 +1,3 @@
+import Dock from "./Dock";
+
+export {Dock}
